Renaming for clarity.
